BUG: seq_len during selection of event-driven idexes is wrong
we randomly insert adverse/helper/unhelper events into
the sequence. Original code gave the entire sequence length
(30, 300, 900) instead of the randomly selected event_length.
When indexes longer that total events is selected,
python inserts it at the end of the event list. This is observed
that majority of a/h/u events are at the end of the sequence.

Changed to use event_length.